# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "username": "user@example.com",
                    "password": "S3gura123",
                    "rolId": 2,
                    "nombre": "María",
                    "apellido": "Pérez",
                    "telefono": "3875789133",
                    "fecha_nacimiento": "1990-05-21",
                    "esta_activo": true,
                    "laboratorioIds": [
                      1,
                      2
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado (sin password en la respuesta)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 10,
                    "username": "user@example.com",
                    "nombre": "María",
                    "apellido": "Pérez",
                    "telefono": "3875789133",
                    "fecha_nacimiento": "1990-05-21",
                    "esta_activo": true,
                    "rol": {
                      "id": 2,
                      "nombre": "Administrador"
                    },
                    "laboratoriosAsignados": [
                      {
                        "id": 1,
                        "laboratorio": {
                          "id": 1,
                          "nombre": "ACME Agro Labs"
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear usuario",
        "tags": [
          "usuarios"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Buscar por username/nombre/apellido/teléfono",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "rolId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campos permitidos: id, username, esta_activo",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [],
                    "meta": {
                      "total": 0,
                      "page": 1,
                      "limit": 10,
                      "pageCount": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar usuarios (paginado y filtrado)",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 10,
                    "username": "user@example.com",
                    "nombre": "María",
                    "apellido": "Pérez",
                    "telefono": "3875789133",
                    "esta_activo": true,
                    "rol": {
                      "id": 2,
                      "nombre": "Administrador"
                    },
                    "laboratoriosAsignados": [
                      {
                        "id": 1,
                        "laboratorio": {
                          "id": 1,
                          "nombre": "ACME Agro Labs"
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener usuario por ID",
        "tags": [
          "usuarios"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "password": "NuevaP4ss",
                    "rolId": 3,
                    "nombre": "Maria",
                    "telefono": "3875000000",
                    "esta_activo": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado (sin password en la respuesta)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar usuario",
        "tags": [
          "usuarios"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar usuario",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/v1/users/{id}/laboratorios": {
      "post": {
        "operationId": "UsersController_addLabs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignLabsDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "laboratorioIds": [
                      1,
                      2,
                      3
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario con laboratorios actualizados"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Asignar laboratorios a un usuario",
        "tags": [
          "usuarios"
        ]
      }
    },
    "/api/v1/users/{id}/laboratorios/{labId}": {
      "delete": {
        "operationId": "UsersController_removeLab",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "labId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "deleted": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Quitar un laboratorio del usuario",
        "tags": [
          "usuarios"
        ]
      }
    }
  },
  "info": {
    "title": "API de Ensayos Agronómicos",
    "description": "Documentación de la API para el sistema de gestión de ensayos",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Email del usuario (único)",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Contraseña en texto plano (será hasheada)",
            "example": "S3gura123",
            "minLength": 6
          },
          "rolId": {
            "type": "number",
            "description": "ID del rol a asignar",
            "example": 2
          },
          "nombre": {
            "type": "object",
            "description": "Nombre",
            "example": "María",
            "maxLength": 100
          },
          "apellido": {
            "type": "object",
            "description": "Apellido",
            "example": "Pérez",
            "maxLength": 100
          },
          "telefono": {
            "type": "string",
            "description": "Teléfono",
            "example": "3875789133",
            "maxLength": 50
          },
          "fecha_nacimiento": {
            "type": "object",
            "description": "Fecha de nacimiento (YYYY-MM-DD)",
            "example": "1990-05-21"
          },
          "esta_activo": {
            "type": "boolean",
            "description": "Usuario activo",
            "example": true,
            "default": true
          },
          "laboratorioIds": {
            "description": "IDs de laboratorios asignados",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "username",
          "password",
          "rolId",
          "telefono"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Email del usuario (único)",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Nueva contraseña (opcional). Si se envía, se vuelve a hashear.",
            "example": "S3gura123",
            "minLength": 6
          },
          "rolId": {
            "type": "number",
            "description": "ID del rol (opcional)",
            "example": 3
          },
          "nombre": {
            "type": "object",
            "description": "Nombre",
            "example": "Juan",
            "maxLength": 100
          },
          "apellido": {
            "type": "object",
            "description": "Apellido",
            "example": "García",
            "maxLength": 100
          },
          "telefono": {
            "type": "string",
            "description": "Teléfono",
            "example": "3875000000",
            "maxLength": 50
          },
          "fecha_nacimiento": {
            "type": "object",
            "description": "Fecha de nacimiento (YYYY-MM-DD)",
            "example": "1992-01-31"
          },
          "esta_activo": {
            "type": "boolean",
            "description": "Usuario activo",
            "example": true,
            "default": true
          },
          "laboratorioIds": {
            "description": "IDs de laboratorios asignados",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "AssignLabsDto": {
        "type": "object",
        "properties": {
          "laboratorioIds": {
            "example": [
              1,
              2,
              3
            ],
            "description": "IDs de laboratorios a asignar",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "laboratorioIds"
        ]
      }
    }
  }
}